Separated host and player views
