+++ frontend/src/App.tsx
@@ -1,5 +1,5 @@
 import './App.css'
-import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
 
 import HomeRoute from './routes/HomeRoute';
 import HostRoute from './routes/HostRoute'
@@ -8,12 +8,14 @@ import PlayerRoute from './routes/PlayerRoute';
 function App() {
   return (
     <BrowserRouter>
-      <Routes>
+      <Routes>        
         <Route path="/" element={<HomeRoute />} />
-        
+
         <Route path="/host/:roomId?" element={<HostRoute />} />
     
         <Route path="/play/:roomId" element={<PlayerRoute />} />
+
+        <Route path="*" element={<Navigate to="/" replace />} />
       </Routes>
     </BrowserRouter>
   );
